Fixing Link Pivot calls

+++ ATSPM/ApplicationCore/Business/LinkPivot/AdjustmentObject.cs
@@ -9,9 +9,9 @@ namespace ATSPM.Application.Business.LinkPivot
         public double UpstreamVolume { get; set; }
         public int LinkNumber { get; set; }
         public double AogTotalBefore { get; set; }
-        public double PAogTotalBefore { get; set; }
+        public int PAogTotalBefore { get; set; }
         public double AogTotalPredicted { get; set; }
-        public double PAogTotalPredicted { get; set; }
+        public int PAogTotalPredicted { get; set; }
         public string ResultChartLocation { get; set; }
         public string DownstreamLocation { get; set; }
         public string DownLocationIdentifier { get; set; }
@@ -21,10 +21,10 @@ namespace ATSPM.Application.Business.LinkPivot
         public double AOGUpstreamPredicted { get; set; }
         public double AOGDownstreamBefore { get; set; }
         public double AOGUpstreamBefore { get; set; }
-        public double PAOGDownstreamPredicted { get; set; }
-        public double PAOGUpstreamPredicted { get; set; }
-        public double PAOGDownstreamBefore { get; set; }
-        public double PAOGUpstreamBefore { get; set; }
+        public int PAOGDownstreamPredicted { get; set; }
+        public int PAOGUpstreamPredicted { get; set; }
+        public int PAOGDownstreamBefore { get; set; }
+        public int PAOGUpstreamBefore { get; set; }
         public int Adjustment { get; set; }
         public int Delta { get; set; }
         public string LocationIdentifier { get; set; }
